requirements.txt
python-telegram-bot==21.4
yt-dlp
python-dotenv
requests
pyopenssl

